Use &Path rather than PathBuf for blockdev paths
We don't strictly need owned paths, since we're not moving, modifying,
or extending them, so we can just take borrowed paths, &Path. The only
substantive change this needs is that in dbus_api.rs create_pool(), instead
of creating a Vec of PathBufs, we create a Vec of &Paths (they still
reference entries in 'devs'.

fixes #53

Signed-off-by: Andy Grover <agrover@redhat.com>