Translate internal errors into DBus-API-specific error values
A StratisError can be of many different varieties, but we want to map all
of these to our fewer choices for DBus API error values.
internal_to_dbus_error() does this in a relatively coarse way that we can
improve on in the future.

Also, use this new function in create_pool().

Signed-off-by: Andy Grover <agrover@redhat.com>