create DeferredAction queue to update the tree outside of dbus method.
The dbus binding doesn't support adding/removing objects while
in a dbus method.  The choice was made to make the tree immutable
while processing a call for an object in that tree.

The author of the binding suggested processing
add/remove operations via a TODO (DeferredAction) queue.

Details can be found at:
https://github.com/diwic/dbus-rs/issues/52

Deferring updates to the tree required the main
message loop to be restructured.  Using the tree.run() method
passed ownership, so the tree became imutable in the loop. So
rather than calling tree.run(), we implement the run code
copied from TreeServer.  This allows us to maintain ownership
of the tree and update it based on the contents of the queue.

Signed-off-by: Todd Gill <tgill@redhat.com>