Add strat_engine
This is the non-simulated Engine implementation. It is incomplete
at this point, but is ready for us to start hacking on it.

Use serde_codegen on strat_engine, to auto-derive Serialize and
Deserialize traits for our data types. This is all rather ugly, I
understand a future Rust feature called "macros 1.1" should make the
ugliness go away.

Signed-off-by: Andy Grover <agrover@redhat.com>