Replace BlockDev::new() with BlockDev::initialize()
If we are creating a pool from multiple blockdevs and there's something
wrong with one of the later ones, we will abort but already have written
Stratis headers to the earlier ones. This new function takes multiple
paths to create blockdevs on at once. This allows it to check all of them
first for min size, is it a blockdev, etc before overwriting the initial
data on any of them.

Rename BlockDev.stratis_id to .pool_id, which is more descriptive.

Signed-off-by: Andy Grover <agrover@redhat.com>