Remove uuids from struct BlockDev
Tracking pool and blockdev uuids in BlockDev is redundant, because there
exist maps of these values to the blockdev elsewhere.

Remove these from BlockDev and pass in to functions like save_state, which
need the values to write them to on-disk metadata.

This also results in changes to return values of setup() and find_all().
setup() now returns a 3-tuple of the values it read off disk, and
find_all() returns nested maps, for convenience.

Signed-off-by: Andy Grover <agrover@redhat.com>