Enrich the return type of Engine::rename_pool().
Now it differentiates between reasons for not taking an action.

D-Bus API now returns an internal error if there was no source pool.

At this point, the only reason for not taking an action is if the mapping
requested was the identity mapping, so it is not yet necessary to enrich
the D-Bus API with something more than bool. If it were, the next step is
to publish codes, kind of like the error codes.

Signed-off-by: mulhern <amulhern@redhat.com>