Move code for batch processing of devices out of BlockDev implementation.
Keep the code as functions in the blockdev module.
Define the auxiliary methods within the methods that they help as they
are not generally useful.
Move the rest of util.rs into blockdev module and remove util.rs.

It is not correct to make batch processing of BlockDevs part of the
BlockDev implementation. The BlockDev implementation should only have
functionality concerning individual BlockDevs.

Currently, BlockDev is not implementing the meager Dev trait, but it is
expected to, and when it does, it will require procedures to implement
those traits.

Distinguish between actual errors during setup of a device and a device
not belonging to Stratis (which should not be considered an error) in setup().

Signed-off-by: mulhern <amulhern@redhat.com>