In real engine's Pool, use device nodes as keys in device maps.
This will be the first thing needed to, e.g., remove devices, which
will typically be specified by paths. This also brings the implementation
into line with the sim engine.

Ultimately, it will be necessary to write a purpose build data-structure
to hold the devices. That can wait until the next time it seems necessary
to look up the BlockDevs by something other than current device node.

The BlockDev currently stores its device node. Removing that is scheduled
for future work along with the purpose build data structure.

Signed-off-by: mulhern <amulhern@redhat.com>