Use string literals instead of defined constants wherever it makes sense.
Move the last needed defined constants into api module.

String literals are by definition static, so it is not necessary to define
as constants to get the correct type for the dbus methods that require
static strings.

Most names are only used in one place, and that is likely to be true for the
foreseeable future for most cases, so the constants do not really serve a
purpose.

Make dependencies between the interface names and the base service name
programmatic.