Add a Redundancy type for engine.
Assume that the set of supported redundancies cannot change dynamically and
will always be the same for both engines.

Change create_pool() so that its parameter is called redundancy, rather
than raid_level.

Within create_pool() convert u16 redundancy value to Redundancy and fail
if conversion fails. The only other alternative is to do the conversion in
dbus_api, which would be putting engine logic in dbus_api.

Pass redundancy as Redundancy type to *Pool::new() methods.

Make redundancy field in Pool private, on the general principle that things
shouldn't be public unless they need to be used that way.

Signed-off-by: mulhern <amulhern@redhat.com>