Add a generic Table data structure
Add a HasName and HasUuid trait to support genericity of Table.

Make Filesystem and Pool extend this trait.

Add ability to rename a pool.

Use Table data structure for Pool and Filesystem maps.

The benefits of the Table are better encapsulation. Look up by either name
or UUID is handled within the data structure rather than by means of separate
data structures or by scanning all items. All operations are of the same
complexity as HashMap's equivalent operations, so they are in theory, O(1).
If it is necessary to adapt the structure to handle very large numbers of
elements, this can be done internally, instead of at every site where it is
used. Lookup by UUID or by name is made explicit in the interface, which
is helpful when reading code that uses the Table structure.

Drop implementation of sime engine's create_snapshot method, it was tricky
to work with during this refactoring.

Signed-off-by: mulhern <amulhern@redhat.com>