Refactor engine::ErrorEnum to implement Error trait
Since all variants of engine::ErrorEnum now have the same associated data,
String, we can pull the string out from within the enum so we can access it
more easily.

The other change here is to implement the standard Error trait for
EngineError, which gives us description(). This is very similar to
get_error_string() but returns a &str.

The changes this causes are mostly mundane, except in dbus_api.rs there is
a small refactoring to take advantage of all EngineErrors now implementing
Error trait.

Also, I think for the non-Io or Nix case I prefer a name other than
"Stratis", "Engine" is slightly better. Since we are changing many lines
for the aforementioned refactoring, do a variant rename at the same time.

Signed-off-by: Andy Grover <agrover@redhat.com>