Move main.rs to bin/stratisd.rs to allow for building stratisd as bin.
This approach follows the recommendations in:
https://doc.rust-lang.org/1.4.0/book/testing.html

Given that there is a natural seperation between the bin stratisd and
the lib engine, splitting them into a binary and a crate fits well.

Other Rust applications have taken different approaches.  For example
cargo uses a technique to rebuild the code used in the "cargo" binary
as a lib crate specifically for testing.  This is done via a separate
build for the integration tests.  This requires the same code to be
compiled twice.  Compiling twice was enough of a disadvantage to
eliminate the alternative.

Note: We should look at the possibility of moving the dbus code into
the bin directory in the future.

Added lib.rs so that the remainder of the application is compiled as a
lib crate.

Moved dbgp! macro to stratisd.rs - it is the only place it is used - we
should probably switch to a logging crate.

Signed-off-by: Todd Gill <tgill@redhat.com>