Update blockdev and metadata for SWDD 0.7.5
Things are now duplicated at the start of the device, instead of start and
end of the device.

We now write a MDA header along with the json, instead of updating the
sig block.

Use EngineResult instead of Result<T, String> in a few places.

Better name structs to match up with SWDD.

Move more disk-manipulation code to metadata by passing in File.

Implement determine_ownership() in terms of setup(), and change it from
taking a buf to taking a File. This lets us eliminate read_pool_uuid() as
well.

Update tests accordingly.

Signed-off-by: Andy Grover <agrover@redhat.com>