Do not implement traits we do not use.
Some of these can be implemented, but we should
think carefully about how this is done.

NewTypeMul should probably never be implemented, as we
are not likely to be able represent Sectors^n, n > 1,
and we are probably not ever going to want that kind of value.

NewTypeDiv is probably wrong, because Sectors/Sectors is a unitless
number. This kind of things works differently with types of numbers.

NewTypeRem is probably correct, but is is unused.

Signed-off-by: mulhern <amulhern@redhat.com>