Make it possible to always cast up (safe), never cast down (uncertain).
Casting up from usize to u64 is safe, but casting down is not.

So, change types of u64 constants that are cast down to usize to usize and
cast up to u64 where necessary.

It is easy to tell by observation where the constant is defined that it fits
into the smaller type.

Constants affected:
* SECTOR_SIZE
* BDA_STATIC_HDR_SIZE
* NUM_PRIMARY_MDA_REGIONS

Signed-off-by: mulhern <amulhern@redhat.com>