Make all fields of MDAHeader that are updated Option type.
last_updated is already Option, because an MDAHeader is not already
necessarily formed with a last_updated value.
data_crc and used, which are in the same situation should be Option
as well.

Also, when loading MDAHeader, use timestamp loaded as None as a signal
that the similar values must be None as well.

Signed-off-by: mulhern <amulhern@redhat.com>