Change destroy_pool to work if block/cache devs are present
It should not be an error to destroy a pool that still contains cache
and/or block devices. Change destroy_pool!() to not error out if so but
instead call the pool destroy() method.

Implement destroy for sim and strat engines. Strat will need more work
once our DM use increases but for the moment we just wipe the
StaticHeaders off block/cache devices.

fixes #234

Signed-off-by: Andy Grover <agrover@redhat.com>