Change DestroyPool() so that it takes an object path as an argument
The new reliance on GetManagedObjects() will allow the client code to
identify the pool-to-be-destroyed by name or UUID and convert that to
the corresponding object path.

The overall direction of this refactoring is to require the client to
identify existing objects by means of their Dbus object paths, either
implicitly, by means of a call on a proxy object, or explicitly, by
passing an object path as an argument.

Signed-off-by: mulhern <amulhern@redhat.com>