Explode api.rs into a few separate private submodules
The separation clarifies the relationships between the various modules
and generally facilitates development.
Moreover, it will be necessary to implement separate interfaces for devices
and possibly for cache devices; that will be more easily handled in
a separate file.

New files are:

One for generic utilities.
One for macros.
One for the filesystem interface.
One for the pool interface.

api.rs still holds code for manager interface.

Signed-off-by: mulhern <amulhern@redhat.com>