Return correct error codes, which depend on the context
If the object path being looked up is the implicit self parameter, then
a lookup failure is an internal error.

On the other hand, if it is some other parameter, then must assume that the
lookup failure is just a client error.

Do not fail if a filesystem object path to be destroyed is missing, because
the goal is to make destruction of filesystems idempotent.

Use macros for common operations.

Signed-off-by: mulhern <amulhern@redhat.com>