Adjust Table implementation
Take into account that name may change, but UUID is constant.
Make T not implement name(), and by implication, not store its own name.
By this means, a variety of bugs of the kind where there are two things
with one name in a Table are prevented. The bugs would not arise by
insertion or removal, which works properly in previous version, but by renaming
of an element within the table.

This preserves theoretical O(1) of all operations. It is possible to
come up with simpler implementations that are O(n), but they do not
avoid the problem of duplicate names on rename any better.

Signed-off-by: mulhern <amulhern@redhat.com>