Add Engine::check()
We will need to call into Engines when we detect an event, and then leave
it up to the Engine to determine the nature of the event, and the correct
response.

Implement a simple method for this, as a starting spot.

Checks are really per-pool, so Engine::check() just calls each pool's
equivalent function.

Signed-off-by: Andy Grover <agrover@redhat.com>