Implement a call to Engine::check() from the stratisd event loop
This is the first time we're calling engine methods not in response to a
DBus request. This has required some code changes since access to the
engine needs to be possible after passing it into dbus_api::connect().

Convert Box<Engine> to Rc<RefCell<Engine>>. This doesn't impact the DBus
code too much because it was already wrapping the engine in an
Rc<RefCell<>>.

Increase poll timeout to 10 seconds.

Signed-off-by: Andy Grover <agrover@redhat.com>