Update tests for new single-device pool capability
I'm not quite sure what was going on with the old test. It was supposed
to fail? In any case, I don't think it should fail (and it doesn't) when
attempting to re-add existing pool members -- they all get filtered out.
Added a TODO in filter_devs because it'd be nice to verify that not
only does the blockdev think it's in this pool, but our data structures
agree. (Otherwise, maybe a udev device-add event was missed?)

Signed-off-by: Andy Grover <agrover@redhat.com>