Do not wipe files backing loopback devices
Previously, we wanted them wiped, but this also served to ensure they were
a particular size. But some distros such as Fedora use in-memory
tmpfs for /tmp, and it will run out of space if multiple Gb-size files are
created there.

Luckily, we don't really need to wipe or fully back our files in tmp. If
we want a file to be a certain size but not write to it, ideally
we would use fallocate(2) to do this. Parts of the file not written-to
don't take up space in the fs, and read as all zeroes. Since nix crate
doesn't yet support fallocate(), a workaround is just to seek() into the
file and then write a little to set the size.

Signed-off-by: Andy Grover <agrover@redhat.com>