Depend on dbus-client-gen, dbus-python-client-gen
Remove all obsoleted code and tests.
Update code and tests for new API of generated classes.

Remove test and coverage targets. The only interesting targets are the
D-Bus based targets but we do not run these in the CI.

No longer use checked_call. We now must rely on the correctness of the
underlying libraries, and we can't test their correctness this way.

Use tuple encoding for redundancy value. Redundancy value is the only
option type in the API.

Remove direct dependency on into-dbus-python. Just have indirect dependency
through dbus-python-client-gen.

Signed-off-by: mulhern <amulhern@redhat.com>