Rewrite find_all() function
Change it so that it returns a mapping of pool UUIDs to a Vec of Devices.

It is premature to instantiate BlockDevs, since the necessary information
in particular the sectors of the blockdev that have been allocated, is not
yet available when the device is first identified as belonging to Stratis.

Instead of loading the entire BDA, just use
StaticHeader::determine_ownership() to find out if the device belongs to
Stratis.

Return an error if a directory entry is an error. This could occur due to
intermittend I/O errors. Better to try to get the device again than to
stand Stratis up with incomplete pools.

Do not read the device node unless it is a block-special device.
Previously, find_all() read all devices to deternmine ownership, which is
crash-prone and inefficient.

Distinguish between errors on opening the device node. For some, skipping
the device is the right course, for others, returning an error seems like
the safest thing.

Resolves: #189.

Signed-off-by: mulhern <amulhern@redhat.com>