Change signature of some metadata reading methods
Have them return Result<Option<T>>. The purpose is to distinguish between
no data there and error when reading the data. This all hinges on the
presence or absence of the Strat Magic number.

Also make some methods that don't have to be public private, for better
encapsulation.

Add a TODO about repairing the static header.

Signed-off-by: mulhern <amulhern@redhat.com>