Remove unused load_state() methods
There are 2, in StratPool and BlockDevMgr. StratPool's is implemented in
terms of the other, as easy to write as to delete. BlockDevMgr's is not
precisely in the right place, it is actually necesary to read the data
prior to the BlockDev's being in a state to be managed. This is what
setup::get_metadata() is for.

Use test that tested BlockDevMgr::load_state() to test
setup::get_metadata() instead.

Signed-off-by: mulhern <amulhern@redhat.com>