Have metadata methods return PoolSave
Previously, they just returned the bytes read. This was unhelpful, because
it left the last step, converting to a struct, to occur at a point that
if it failed, it was not possible to try another device. Now, the
method just tries a next device on failure.

Remove a now obsoleted test.

Signed-off-by: mulhern <amulhern@redhat.com>