Combine RangeAllocator::new() with RangeAllocator::new_with_used()
Keep the name new(), get rid of new_with_used().

It is too easy to pass the empty vec to mean "nothing used" to make it
worthwhile to have distinct methods. The name new_with_used() occasionally
sounds wierdly paradoxical, or at least awkward, easier if it is not
there.

Add a test.

Signed-off-by: mulhern <amulhern@redhat.com>