Get rid of an unwrap() in source code
Treat failure to obtain the devnode() as an error in initialize().

We are trying to eliminate unwrap()'s in source.
An unwrap() should be converted into an expect() with a string explaining
why it must always succeed or it should be treated as an error if it
might not succeed. If we must have an unwrap() in the source it should
be accompanied by a TODO, explaining how it will be fixed in future.

Signed-off-by: mulhern <amulhern@redhat.com>