Return Segments from BlockDev methods
We already make use of devicemapper-rs types.
Our other methods always want Segments.

Also, unpublicise dev field in BlockDev. It was only public in order to
use a method that was altered in this commit.

Get rid of unused dstr() method. If we need it again, we can add it.

Signed-off-by: mulhern <amulhern@redhat.com>