Have various types implement Isomorphism
Return the segments that back the mdv device.
These are needed in order to construct metadata.

This brings us to the point where ever field specified in the design doc
that can be saved under the current state of the implementation is saved.

Complicate to_save() method of Isomorphism trait by allowing it to error.
It can, due to the necessary lookup from device to our blockdev UUID.

Signed-off-by: mulhern <amulhern@redhat.com>