Change real space backing thin devs to 768 MiB
Updated the initial size of real space allocated to back
a thin pool to 768 MiB.  Stratis currently gives thin devices a
virtual size of 1 TiB.  The XFS filesystem consumes ~547 MiB for
meta data when provisioned over a 1 TiB device.  If the
INITIAL_DATA_SIZE is only 512 MiB, Stratis will need to extend the
pool as the filesystem is writing the meta data.  Thin pool
expansion is not yet in place.  This update will allow provisioning
XFS without running out of real space in the pool.

Signed-off-by: Todd Gill <tgill@redhat.com>