Remove setup_thinpooldev
We can't have a single function because we need different behavior for
initialize and setup. initialize wants to wipe the meta device and use a
default data block size. setup wants to *not* wipe the meta device, get
the data block size from the metadata, and will soon catch NeedsCheck from
ThinPoolDev::setup so it can call thin_repair if needed.

Signed-off-by: Andy Grover <agrover@redhat.com>