Improve BlockDevMgr::save_state()
Get rid of the unwrap(). Consider the result a success if the metadata is
written somewhere. Obviously, for large pools, it would be desirable to
write to more than one blockdev, but if at least one was written to, it
should be possible to find the most up-to-date data next time it is needed,
or on setup.

Signed-off-by: mulhern <amulhern@redhat.com>