Do not verify the crc of the variable length metadata on initial load
If we do this, then loading the BDA will fail if either of the region's
variable length metadata appears invalid. This is a problem, because during
setup, if there is an error reading the BDA, setup terminates.
get_metadata() should continue if the only problem with the BDA is
the variable length metadata on one region.

In a subsequent stage in get_metadata() the metadata is read and it is
checked whether it is valid. In that situation, if it is invalid, that
device is skipped, as there may be other devices in the pool with valid
variable length metadata that can be used to construct the pool.

Signed-off-by: mulhern <amulhern@redhat.com>