Add --verbose flag to clippy target
Since it is only run on nightly, it may allow us to spot regressions
in nightly (or clippy) that we should report to the appropriate people.
Generally, verbose is annoying, because it is more verbose, but not,
usually, more helpful to me. In this case, though, it is probably worth
it.

Signed-off-by: mulhern <amulhern@redhat.com>