Enrich return type of BlockDevMgr::alloc_space
Have it return the number of sectors allocated.
Update call sites appropriately.
Revise StratPool::extend_data() so that it does not fail if it is at least
able to allocate some sectors, even if not all requested.
Add an assertion.

This change makes alloc_space() more informative but also trickier to use.
If calling code ever decides not to use the allocated space, then it must
return it, otherwise we'll be leaking sectors.

Signed-off-by: mulhern <amulhern@redhat.com>