Rewrite the building of bdas vec using collect() and and_then()
This is a style change but also a behavior change: keep going if file
open() or BDA::load fails, instead of exiting with a failure. Yes, the
failed device may have contained the latest metadata, but if it can't be
read then it is akin to a missing device, and we should still proceed.

This seems in line with the later code that continues if the variable
metadata can't be read.

Signed-off-by: Andy Grover <agrover@redhat.com>