Make get_blockdevs() consistent with get_metadata()
It now ignores devices without a BDA or devices which do not belong to the
specified pool just like get_metadata() does.

This fixes a bug where the method could fail if it was given a devnode that
did not belong to the pool. It would fail in the last stage, when it found that
it had constructed a blockdev that didn't match the metadata.

Signed-off-by: mulhern <amulhern@redhat.com>