Distinguish between getting (im)mutable pools and filesystems
We were using functions that got mutable pools and filesystems in all
cases, but that will cause difficulties in future.

Also, we get to remove a few mut annotations by using the non-mut versions
of the functions, which is generally beneficial.

In D-Bus API, a mutable pool is always required, so just change
get_pool macro to get_mut_pool macro.

Signed-off-by: mulhern <amulhern@redhat.com>