Destroy the blockdevs if it is impossible to create the ThinPool
Previously, we were destroying them only if there was not enough space.

This corrects a bug, as if we can't make the pool we should release the
blockdevs. However, it may also increase the possibility of confusion if
the blockdevs are erased but some devicemapper devices constructed from
the blockdevs are still hanging around.

Also, make initial_size() method private as it is no longer needed in
StratPool.

Signed-off-by: mulhern <amulhern@redhat.com>