Enable test_thinpool_device to use one blockdev via BlockDevMgr
Using BlockDevMgr will let this function be less tied to the dimensions
of the blockdevs it is given.

One issue this resolves is that if the initial blockdev was very large,
before it was used as the metadata blockdev it was completely wiped, which
could take a long long time. Now it tries to allocate 16 MiB for metadata
device and 1 GiB for data device. This will let it use just one blockdev,
if it is large enough.