Increase descriptiveness of test device number requests
Make the number given to test_with_spec() not just a number, but wrap it
in an enum to describe what the number means: The test needs exactly this
many, or at least this many? There's also a Range enum to limit max if
needed.

Each test needs to specify the minimum number of devices it needs, so
that we can error out if that number is not available. But, the max number
should usually depend on the number actually available. (Use Range if not.)

Use Exactly to keep test behavior identical to status quo.

Signed-off-by: Andy Grover <agrover@redhat.com>