suspend running loopback tests on Travis
Suspend the loopback tests until we either get access to newer xfsprogs
on the Travis VM or are able to run inside a container.

Stratis requires xfsprogs 4.2.0 or greater.  The -U <uuid> feature is
used to keep the Stratis UUID for the filesystem in sync with the UUID
used by XFS.

Signed-off-by: Todd Gill <tgill@redhat.com>