dbus snapshot interface /  specify a name for a snapshot
Add dbus interface to snapshot a filesystem via the pool

Once created a snapshot appears to the user as the same as any filesystem.
The snapshot shares common blocks with the origin, but the user of the API
doesn't need to be concerned with the connection.

Updates to allow the caller to specify the Stratis name for the new
filesystem created as a result of snapshot.

Signed-off-by: Todd Gill <tgill@redhat.com>