Remove blockdevs from D-Bus when the pool is destroyed
Add a "children" HashSet to OPContext. This lets us keep a list of a pool's
blockdev paths, so that we can easily remove them if the pool is destroyed.
Since get_data() returns an immutable reference, wrap the hashset in a
RefCell to get around this.

Change places in the code that create DBus objects for pool and blockdevs
to also update the above hashset.

Change create_dbus_pool() to return ObjectPath instead of Path. This is
needed because we need to add the paths of its blockdevs to its data, which
is not possible if it's already been placed on the action list.

fixes #626

Signed-off-by: Andy Grover <agrover@redhat.com>