dbus api: Handle deferred at start up
When api::connect is executed the code iterates over the engine processing
all the objects.  It makes calls to create_dbus_pool, create_dbus_filesystem,
and create dbus_block_dev.  The problem is all of these functions add the
changes to the "action" queue.  Thus when then the first client does any
method call it will appear the system is empty until after we process
the deferred event queue after handling the initial request.

This change moves the deferred processing into a separate function and
calls it in api::connect before returning, so that the tree has all the
queued objects applied to it.

Note: Updated to include changes from Grover which pass the actions
instead of the entire dbus context.

Signed-off-by: Tony Asleson <tasleson@redhat.com>