Make process_deferred_actions take a simple ActionQueue argument
The advantages of this are that process_deferred_actions has a cleaner
interface, and the borrow_mut()'s are all in the places that they occur
rather than hidden in process_deferred_actions. This makes it a bit easier
to track them and to make sure that they do not conflict with other
borrows.

Signed-off-by: mulhern <amulhern@redhat.com>