Better handle panics in unit tests
Before we execute the unit test we will do a sanity check on the
test environment which will panic and end that unit test if bad.  If
the environment is good we will run the test capturing the result
and then clean up capturing its result.  We will then ensure that
the test and the post clean ups were successful as well.  This
approach removes the potential of a panic during a panic and allowing
us to know when we encounter an issue during unit tests.

Note: This change incorporates a number of suggestions/improvements
from Mulhern.

Signed-off-by: Tony Asleson <tasleson@redhat.com>
Signed-off-by: mulhern <amulhern@redhat.com>