Use common function for CLI calls
Create a common function which captures stdout & stderr so that we
never see it get spewed out to the journal unless we want to have it
in a logged error message.  This also has a nice side effect of
removing copy & pasted code.

Note: This is a follow on to changes introduced with 94dbe87.

Signed-off-by: Tony Asleson <tasleson@redhat.com>