Reimplement Table
Use a Name->Uuid map as well as a Uuid->(Name, T) map. Name is included in
the Value of the second map in order to play nice with Rust and
implementing mutable iterators, see
https://users.rust-lang.org/t/itermut-and-lifetime-issue/14117 for more.
This is made somewhat less costly because Name is refcounted, see prev
commit.

The API changed slightly because insert now can only ever return one
previous entry, so insert now returns Option<T> instead of Vec<T>. Also,
remove empty() and alter code that used it.

The primary issue with this commit is that using the mutable iterator, it's
possible to change the internal name and uuid values of T, which would
introduce a bug through inconsistency with the keys. The potential
solutions are: Don't implement IterMut (not an option since we
need to use it in a number of places); or just don't store name & Uuid
in T. The following commits explore the latter option, but we could just
apply commits up to this point if desired.

Alphabetize use decls in structures.rs.

Signed-off-by: Andy Grover <agrover@redhat.com>