Improve check()
Change check to return a bool to communicate if a change has been made that
requires us to save a new version of the on-disk metadata.

Add an assert in event_on(). We will in the future possibly get events for
other DM devices in the pool, but right now we should not.

fixes #681

Signed-off-by: Andy Grover <agrover@redhat.com>