Handle udev 'add' events
Block devices can be delayed at boot or be systematically added
after the system is up and running.  Handle these udev events by
evaluating block devices when they are added, to see if they have
a valid stratis signature.  If they do we will attempt to assemble
it into a complete pool.

Signed-off-by: Tony Asleson <tasleson@redhat.com>