Introduce cachedevs into the D-Bus layer
1. Add a new BlockDevTier property to dbus blockdev interface

This will allow to distinguish between cache and data devs.
Extend Pool::get_blockdev and Pool::get_mut_blockdev so that they return
the tier of the located blockdev.

This approach keeps the D-Bus and engine layer separate.
The alternative approach, which extends the Pool interface with additional
methods, entangles the two layers more, as the D-Bus layer has to put more
meaning into which methods it calls.

2. Add a BlockDevTier parameter to Pool::add_blockdevs

Return an error if BlockDevTier::Cache is passed, because no code yet
exists to add the cache blockdevs. In the sim engine, actually store the
cachedevs.

3. Add an AddCacheDevs method to the D-Bus API

Change AddDevs to add AddDataDevs for easier distinguishing of the two.

Share the implementation the methods add_cachedevs and add_blockdevs in the
API. This may be a temporary thing.

The idea of having AddCacheDevs be a separate method in the D-Bus API is
that at some point the parameters of AddCacheDevs and AddBlockDevs might
diverge.  On the other hand, they haven't yet, so it is somewhat reasonable
to share the internal implementation. It is also very cheap in programmer
time, so it is a reasonable way to start.

Signed-off-by: mulhern <amulhern@redhat.com>