Pass last_update_time as a parameter in BlockDevMgr::new
Previously, it was set unconditionally.

If the parameter is not passed then the method is restricted in how and
when during execution it can be invoked. If we choose not to pass the
parameter, then we should document BlockDevMgr::new() a whole lot better.
It is easier to pass the parameter.

Signed-off-by: mulhern <amulhern@redhat.com>