Refactor so that backstore is in separate module
Extend backstore so that it presents a DM linear device to the upper layers.

This is an initial step in the implementation of the cache tier.
The intention is that it will be possible to swap dynamically between a
linear device and a cache device. This justifies the current situation
where the Backstore struct has only one field, the DataTier.

This necessitates a change in the metadata which is expected to be
temporary: https://github.com/stratis-storage/stratisd/issues/762.
The thinpool module now allocates all space from the backstore rather
than directly from the blockdev manager.

It adds an additional check for coherence of metadata: the amount allocated
to the devices in the stratis pool must equal the amount allocated from
the backstore.

It adds a new kind of standardized stratis names, those which describe the
DM devices in the cache tier.

It redistributes the management of thinids to a separate file.

This also redistributes some setup and cleanup code among the newly created
module and other modules.

It makes one test less stringent than it was formerly, in that it does not
specify the type of error that must be returned when instantiating an empty
pool, just that there must be an error.

It removes some code from one test. This test was added in a previous PR
to verify the correctness of a bug fix, but is now removed. The errors
that it detected should now be detected by failures in assertions in the
code itself. See: https://github.com/stratis-storage/stratisd/pull/749.

It makes get_coalesced_segments a little more generally useful; previously
it relied on a thinpool invariant. It changes its name to
coalesce_blkdevsegs.

It adds no new tests, but passes all existing tests.

Signed-off-by: mulhern <amulhern@redhat.com>