api proxy
The goal is to try and limit cfg directives sprinkled everywhere. It's
bad enough as it is!

change connect to take a passed-in connection

use proxy in thinpool check()

pass in ApiProxy to StratEngine initialize()

Make engine.pools() return a mut ref, we need it to call set_dbus_path().