When constructing a StratBlockDev, ignore additional sectors
The blockdev may have become larger for some reason since the last time
stratis looked at it. If it did, just pretend the extra sectors aren't there.

In DataTier::setup() treat any additional space available from the
blockdevmgr as an error. All sectors are always allocated immediately to
the DM device. If there are extra sectors, that means probable metadata
corruption.

Remove StratBlockDev::recorded_size(). Since it is unaware of any extra
blocks it might have available to it, the value of recorded_size() and of
current_capacity() must always be the same.

Signed-off-by: mulhern <amulhern@redhat.com>