Implement support for cachedevs in strat engine
This involves numerous changes in the implementation of Backstore and
its component devices. It also requires changing the thinpool module so
that it is possible to retarget the various DM devices in the thinpool
in case a cache device is created after the thinpool DM devices have
already been mapped to the cache's origin device.

Changes are:

* Extend definition of Backstore variable length metadata. Add information
about its cache devs and also about its allocation status.
* Within the thinpool module add numerous suspend and resume methods to
suspend and resume the various DM devices that make up the thinpool
components.
* Add the ability to set the table of the linear DM device which backs the
metadata vol. Note that this is necessary for other actions which might
have to be done as a result of stratisd self-monitoring, currently done
by the check method, but that that has not been implemented yet.
* Modify StratPool::add_blockdevs so that it suspends the thinpool before
adding blockdevs to the backstore and so that it sets the device for
all the thinpool components. If the device has not changed this is a no-op,
but if the device has changed, as will happen when a cache is built, all
tables belonging to DM devices in thinpool need to be reloaded.
* Implement the set_device() method in ThinPool.
* Numerous changes in backstore module:
  - Take the DM device out of the DataTier. It has to be passed around so
    much that giving the DataTier ownership is a bad idea. This requires
    modifying DataTier::setup() and DataTier::new() so that they return
    the constructed DM device separately, i.e., not as a component of the
    DataTier. This also requires that DataTier::capacity() gets its info
    from its segments field.
  - No longer track allocation (with next) in the DataTier. It must be
    tracked at the Backstore level.
  - Modify DataTier::add() so that it takes either a cache or a linear
    device as a parameter. This is necessary, because once the cache is
    established, the origin device of the cache tier must be extended
    via cache methods.
  - Introduce a new struct, CacheTier. This struct is analogous to
    DataTier but handles the segments associated with cache specific
    DM devices.
  - Modify the Backstore adding additional fields to track a cache tier
    and the mutually exclusive DM devices, cache and linear.
  - Extend Backstore::add_blockdevs() so that it manages addition of more
    cachedevs and initial construction of cache dev if necessary.
  - Perform allocation in Backstore.
* Add some tests

Signed-off-by: mulhern <amulhern@redhat.com>